startup-kpi-dashboard/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-dateutil>=2.8.0
